Add basic instruments interface & note queue
We can easily implement new instruments that have a state & whose outputs
are a function of time. The server plays these sounds when their start time
has (almost) passed.

I think I should rename Instrument to Note. I don't need to think about
efficiency; if Notes are re-used (like, buffers are not alloced &
dealloced), this is dealt with some Allocator.

For an example of the necessity of the state, we can think of reverb.

What I don't know yet is how to elegantly and efficiently chain filters,
merge sounds, deal with stereo. Elegantly: find a nice syntax that
allows me to create new sounds quickly. FP seems better for that; maybe
I need a DSL. Efficiently: if there is a chain of filters that give
buffers b1→b2→b3 and b1→b2→b3', I want to compute b1 and b2 only
once.

One thing I need to do soon is to log when a note is added to the server.
Each Note can have a simple representation that we can serialize easily.